[fix:34] files organization Co-authored-by: Jose Eduardo <jevrprado@gmail.com> Co-authored-by: @annacbrandao <annacebrandao@gmail.com>
