+++ src/Pages/Protected/Benefit/Benefits/index.jsx
@@ -1,10 +1,10 @@
 //import { useEffect, useState } from "react";
-import { useAuth } from "../../../Context/auth";
+import { useAuth } from "../../../../Context/auth";
 import { useNavigate } from "react-router-dom";
 //import { APIUsers } from "../../../Services/BaseService";
-import SecondaryButton from "../../../Components/SecondaryButton";
-import sindpol_logo from "../../../assets/sindpol-logo.png";
-import sentinela_logo from "../../../assets/sentinela-logo.png";
+import SecondaryButton from "../../../../Components/SecondaryButton";
+import sindpol_logo from "../../../../assets/sindpol-logo.png";
+import sentinela_logo from "../../../../assets/sentinela-logo.png";
 
 import "./index.css";
