+++ src/Pages/Protected/Benefit/BenefitsList/index.jsx
@@ -1,9 +1,9 @@
 import { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
-import PrimaryButton from "../../../Components/PrimaryButton";
-import SecondaryButton from "../../../Components/SecondaryButton";
-import FieldText from "../../../Components/FieldText";
-import { APIBenefits } from "../../../Services/BaseService";
+import PrimaryButton from "../../../../Components/PrimaryButton";
+import SecondaryButton from "../../../../Components/SecondaryButton";
+import FieldText from "../../../../Components/FieldText";
+import { APIBenefits } from "../../../../Services/BaseService";
 import List from "@mui/material/List";
 import ListItem from "@mui/material/ListItem";
 import ListItemButton from "@mui/material/ListItemButton";
